# backend/docs/swagger.yaml
basePath: /
definitions:
  controllers.CompareResult:
    properties:
      fixed:
        items:
          $ref: '#/definitions/models.Finding'
        type: array
      from:
        $ref: '#/definitions/models.ApplicationVersion'
      new:
        items:
          $ref: '#/definitions/models.Finding'
        type: array
      stillPresent:
        items:
          $ref: '#/definitions/models.Finding'
        type: array
      to:
        $ref: '#/definitions/models.ApplicationVersion'
    type: object
...
  gin.H:
    additionalProperties: {}
    type: object
...
  models.ApplicationVersion:
    properties:
      applicationId:
        type: integer
      branch:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      isDefault:
        type: boolean
      name:
        type: string
      tag:
        type: string
      updatedAt:
        type: string
    type: object
...
  models.Finding:
    properties:
      applicationVersion:
        $ref: '#/definitions/models.ApplicationVersion'
      applicationVersionId:
        type: integer
      assignedTo:
        type: integer
      assignedToUser:
        $ref: '#/definitions/models.User'
      createdAt:
        type: string
      cweId:
        type: string
      dedupeHash:
        type: string
      description:
        type: string
      dueDate:
        type: string
      epssScore:
        type: number
      filePath:
        type: string
      fixedAt:
        type: string
      id:
        type: integer
      lineEnd:
        type: integer
      lineStart:
        type: integer
      remediation:
        type: string
      reviewedBy:
        type: integer
      reviewedByUser:
        $ref: '#/definitions/models.User'
      riskScore:
        type: number
      ruleId:
        type: string
      scanId:
        type: integer
      scannerType:
        $ref: '#/definitions/models.ScannerType'
      scannerTypeId:
        type: integer
      severity:
        type: string
      status:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  models.ScannerType:
    properties:
      createdAt:
        type: string
      description:
        type: string
      enabled:
        type: boolean
      id:
        type: integer
      name:
        type: string
      parser:
        type: string
      updatedAt:
        type: string
    type: object
...
  models.User:
    properties:
      avatarUrl:
        type: string
      banned:
        type: boolean
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      oauthId:
        type: string
      oauthProvider:
        type: string
      role:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
  models.Webhook:
    properties:
      applicationId:
        type: integer
      createdAt:
        type: string
      events:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      secret:
        type: string
      updatedAt:
        type: string
      url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@servasec.local
    name: API Support
  description: Security scanner management platform API
  title: Servasec API
  version: "1.0"
paths:
...
  /applications/{id}/versions:
    get:
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of versions
          schema:
            items:
              $ref: '#/definitions/models.ApplicationVersion'
            type: array
      summary: List versions
      tags:
      - Applications
    post:
      consumes:
      - application/json
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Version details
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created version
          schema:
            $ref: '#/definitions/models.ApplicationVersion'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: Version with this name already exists
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create version
      tags:
      - Applications
  /applications/{id}/versions/{versionId}:
    delete:
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Version ID
        in: path
        name: versionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "404":
          description: Version not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete version
      tags:
      - Applications
    get:
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Version ID
        in: path
        name: versionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Version details with scans
          schema:
            $ref: '#/definitions/models.ApplicationVersion'
        "404":
          description: Version not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get version
      tags:
      - Applications
    patch:
      consumes:
      - application/json
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Version ID
        in: path
        name: versionId
        required: true
        type: string
      - description: Fields to update
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Updated version
          schema:
            $ref: '#/definitions/models.ApplicationVersion'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Version not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update version
      tags:
      - Applications
  /applications/{id}/versions/{versionId}/findings:
    get:
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Version ID
        in: path
        name: versionId
        required: true
        type: string
      - description: Filter by severity
        in: query
        name: severity
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of findings
          schema:
            items:
              $ref: '#/definitions/models.Finding'
            type: array
        "404":
          description: Version not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: List version findings
      tags:
      - Applications
  /applications/{id}/versions/compare:
    get:
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Source version ID or name
        in: query
        name: from
        required: true
        type: string
      - description: Target version ID or name
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Comparison result
          schema:
            $ref: '#/definitions/controllers.CompareResult'
        "400":
          description: Missing from/to params
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Version not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Compare versions
      tags:
      - Applications
  /applications/{id}/webhooks:
    get:
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of webhooks
          schema:
            items:
              $ref: '#/definitions/models.Webhook'
            type: array
      summary: List webhooks
      tags:
      - Webhooks
    post:
      consumes:
      - application/json
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Webhook details
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created webhook
          schema:
            $ref: '#/definitions/models.Webhook'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create webhook
      tags:
      - Webhooks
  /applications/{id}/webhooks/{webhookId}:
    delete:
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Webhook ID
        in: path
        name: webhookId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "404":
          description: Webhook not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete webhook
      tags:
      - Webhooks
...
securityDefinitions:
  ApiKeyAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"